Correccion en repositorio

+++ API/DTOS/ContactDTO.cs
@@ -2,7 +2,7 @@ namespace DTOS
 {
     public class ContactDTO
     {
-        public int Id{set;get;}
+        public long Id{set;get;}
         public string CompleteName {set;get;}
         public string Mail{set;get;}
         public string Message{set;get;}
