+++ API/Models/ContactEntity.cs
@@ -4,7 +4,7 @@ namespace Models
 {
     public class ContactEntity
     {
-        public int Id {set;get;}
+        public long Id {set;get;}
         public string CompleteName {set;get;}
         public string Phone {set;get;}
         public string Mail{set;get;}
